Refactor the rest of the tests
