+++ test/test-data/invalidate-field.ts
@@ -1,0 +1,86 @@
+import { OneTest } from "./one-test";
+import { invalidateField } from "../../src";
+
+const testObj1 = { id: "obj1", name: "foo" };
+const testObj2 = { id: "obj2", names: ["foo", "bar"] };
+const testObj4 = { id: "obj4", ["names(id:1)"]: ["foo", "bar"] };
+
+export const invalidateFieldTestData: ReadonlyArray<OneTest> = [
+  {
+    name: "invalidateFieldRecursive with shallow data",
+    patches: [invalidateField<typeof testObj2>("obj2", "names", true)],
+    cacheBefore: { obj2: testObj2 },
+    cacheAfter: { obj2: testObj2 },
+    staleBefore: {},
+    staleAfter: { obj2: { names: true } }
+  },
+  {
+    name: "invalidateFieldWithParamsRecursive",
+    patches: [invalidateField<any>("obj4", "names", true)],
+    cacheBefore: { obj4: testObj4 },
+    cacheAfter: { obj4: testObj4 },
+    staleBefore: {},
+    staleAfter: { obj4: { ["names(id:1)"]: true } }
+  },
+  {
+    name: "invalidateFieldShallow with deep data",
+    patches: [invalidateField<any>("myid", "prop", false)],
+    cacheBefore: {
+      myid: { prop: "obj:1" },
+      ["obj:1"]: { id: "1", news: ["NewsItemType:1", "NewsItemType:2"] },
+      ["NewsItemType:1"]: { id: 1, header: "olle" },
+      ["NewsItemType:2"]: { id: 2, header: "olle2" }
+    },
+    cacheAfter: {
+      myid: { prop: "obj:1" },
+      ["obj:1"]: { id: "1", news: ["NewsItemType:1", "NewsItemType:2"] },
+      ["NewsItemType:1"]: { id: 1, header: "olle" },
+      ["NewsItemType:2"]: { id: 2, header: "olle2" }
+    },
+    staleBefore: {},
+    staleAfter: {
+      myid: { prop: true }
+    }
+  },
+  {
+    name: "invalidateFieldRecursive with deep data",
+    patches: [invalidateField<any>("myid", "prop", true)],
+    cacheBefore: {
+      myid: { prop: "obj:1" },
+      ["obj:1"]: { id: "1", news: ["NewsItemType:1", "NewsItemType:2"] },
+      ["NewsItemType:1"]: { id: 1, header: "olle" },
+      ["NewsItemType:2"]: { id: 2, header: "olle2" }
+    },
+    cacheAfter: {
+      myid: { prop: "obj:1" },
+      ["obj:1"]: { id: "1", news: ["NewsItemType:1", "NewsItemType:2"] },
+      ["NewsItemType:1"]: { id: 1, header: "olle" },
+      ["NewsItemType:2"]: { id: 2, header: "olle2" }
+    },
+    staleBefore: {},
+    staleAfter: {
+      myid: { prop: true },
+      ["obj:1"]: { id: true, news: true },
+      ["NewsItemType:1"]: { id: true, header: true },
+      ["NewsItemType:2"]: { id: true, header: true }
+    }
+  },
+  {
+    name:
+      "invalidateFieldRecursive should do nothing if entity is missing in cache",
+    patches: [invalidateField<typeof testObj2>("notobj2", "names", true)],
+    cacheBefore: { obj2: testObj2 },
+    cacheAfter: { obj2: testObj2 },
+    staleBefore: {},
+    staleAfter: {}
+  },
+  {
+    name:
+      "invalidateFieldRecursive should do nothing if field is missing in cache",
+    patches: [invalidateField<typeof testObj2>("obj2", "names", true)],
+    cacheBefore: { obj2: testObj1 },
+    cacheAfter: { obj2: testObj1 },
+    staleBefore: {},
+    staleAfter: {}
+  }
+];
